requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
mailjet-rest==1.3.4
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.8
requests==2.27.1
smart-open==5.2.1
spacy==3.3.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0.tar.gz#egg=en_core_web_sm
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
thinc==8.0.15
tqdm==4.64.0
typer==0.4.1
typing-extensions==4.2.0
urllib3==1.26.9
visitor==0.1.3
wasabi==0.9.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
